massive update!
so huge update. created a commentaries section that works very similarly to the articles section. for sure could be some optimizations and will take a ton of time to get all the content up. all in good time! also fixed the issue of having to load font in every single page.